Draft of left sidebar in search
* Added get_structure func to utils
   - Merges structures to acquire full list of parts
   - Moved from homepage.py
   - Should core do this instead?
* Added 'left_sidebar_template' parameter to context
   - Problem: Default left sidebar view not correct for search
   - Solution: Allow different templates for different data types
   - Downside: All views must specify a template :(
