+++ regulations/templates/regulations/partials/left_sidebar_views/subpart_view.html
@@ -1,0 +1,13 @@
+{% for item in toc.children %}
+    {% if item.type == "section" %}
+        {% include "regulations/partials/toc_section.html" with extra_class="orphan" %}
+    {% elif item.type == "subpart" %}
+        {% with item_subpart="Subpart-"|add:item.identifier.0 %}
+            {% if subpart and item_subpart == subpart %}
+                {% include "regulations/partials/toc_subpart.html" with collapse_state="expanded" %}
+            {% else %}
+                {% include "regulations/partials/toc_subpart.html" with collapse_state="collapsed" %}
+            {% endif %}
+        {% endwith %}
+    {% endif %}
+{% endfor %}
