+++ regulations/views/reader.py
@@ -39,6 +39,7 @@ class ReaderView(TableOfContentsMixin, CitationContextMixin, TemplateView):
             'toc':          toc,
             'parts':        parts,
             'versions':     versions,
+            'left_sidebar_content': 'regulations/partials/left_sidebar_views/subpart_view.html',
         }
 
         return {**context, **c}
