+++ regulations/views/search.py
@@ -1,6 +1,11 @@
+from datetime import date
+
 from django.views.generic.base import TemplateView
 
 from regulations.generator.api_reader import ApiReader
+from .utils import get_structure
+
+client = ApiReader()
 
 
 class SearchView(TemplateView):
@@ -10,9 +15,17 @@ class SearchView(TemplateView):
     def get_context_data(self, **kwargs):
         context = super().get_context_data(**kwargs)
         results = get_data(self.request.GET.get("q"))
-        context['results'] = results
-        return {**context, **self.request.GET.dict()}
+        today = date.today()
+        parts = client.effective_parts(today)
+        structure = get_structure(parts)
+        c = {
+            'parts': parts,
+            'toc': structure,
+            'results': results,
+            'left_sidebar_content': 'regulations/partials/left_sidebar_views/title_view.html',
+        }
+        return {**context, **c, **self.request.GET.dict()}
 
 
 def get_data(query):
-    return ApiReader().search(query)
+    return client.search(query)
