+++ regulations/views/utils.py
@@ -23,3 +23,21 @@ def find_subpart(section, node, subpart=None):
             if value is not None:
                 break
     return value
+
+
+def different(one, two):
+    return one['identifier'] != two['identifier']
+
+
+def merge_children(one, two):
+    if different(one[len(one)-1], two):
+        one.append(two)
+        return
+    merge_children(one[len(one)-1]['children'], two['children'][0])
+
+
+def get_structure(parts):
+    structure = [parts[0]['structure']]
+    for part in parts[1:]:
+        merge_children(structure, part['structure'])
+    return structure
